NetworkSpeed component refactor into multiple components
